% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.R
\name{foldercloud}
\alias{foldercloud}
\title{foldercloud}
\usage{
foldercloud(folder, cloudname = NULL, subfolders = TRUE,
  max.words = 500, scale = c(3.8, 0.29), min.freq = 4, width = 16,
  height = 16, units = "cm", res = 600, exclude = c("journal"),
  ...)
}
\arguments{
\item{folder}{a string. The path to a folder}

\item{cloudname}{a string. The name of the destination file. Include document type. Default is "wordcloud.png".}

\item{subfolders}{a logical. If TRUE, the algorithm will find all pdfs in all subfolders of the target path}

\item{max.words}{a numeric. Maximum number of words to be plotted. least frequent terms dropped}

\item{scale}{a numeric vector. Scale indicates how minimum and maximum word size.}

\item{min.freq}{a numeric. Words with frequency below min.freq will not be plotted}

\item{width}{a numeric. How many units wide to make the final plot.}

\item{height}{a numeric. How many units high to make the final plot.}

\item{units}{a string. The units in which height and width are given. Can be px (pixels), in (inches), cm(default) or mm.}

\item{res}{a numeric. The nominal resolution in ppi (pixels per inch) which will be recorded in the bitmap file. Default is 600ppi.}

\item{exclude}{a charcater vector. Words in this vector will be excluded from the final plot.}

\item{...}{additional arguments may be passed to the word cloud here.}
}
\value{
A list containing a word frequency table, and a wordcloud
}
\description{
takes in the path of a folder containing pdf. Performs frequency analysis and produces a wordcloud.
}
