% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.R
\name{prevent_duplicates}
\alias{prevent_duplicates}
\title{prevent_duplicates}
\usage{
prevent_duplicates(path = NULL)
}
\arguments{
\item{path}{character UNIX path}
}
\description{
If a path is going to be overwritten, prevent that.
}
