% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_pdfs.R
\name{search_pdfs}
\alias{search_pdfs}
\title{search_pdfs}
\usage{
search_pdfs(folder, pattern, exec = T, ...)
}
\arguments{
\item{folder}{the path to a folder}

\item{pattern}{regex search pattern}

\item{exec}{a bool. If true, the pdfs will be openeded (windows only)}

\item{...}{additional arguments to grepl}
}
\description{
Search through pdfs for target words
}
